// src/interfaces/Options.ts
interface SchemaDumpOptions {
    /**
     * True to include autoincrement values in schema, false otherwise.
     * Defaults to true.
     */
    autoIncrement?: boolean;
    /**
     * True to include engine values in schema, false otherwise.
     * Defaults to true.
     */
    engine?: boolean;
    /**
     * True to run a sql formatter over the output, false otherwise.
     * Defaults to true.
     */
    format?: boolean;
    /**
     * Options for table dumps
     */
    table?: {
        /**
         * Guard create table calls with an "IF NOT EXIST"
         * Defaults to true.
         */
        ifNotExist?: boolean;
        /**
         * Drop tables before creation (overrides `ifNotExist`).
         * Defaults to false.
         */
        dropIfExist?: boolean;
        /**
         * Include the `DEFAULT CHARSET = x` at the end of the table definition
         * Set to true to include the value form the DB.
         * Set to false to exclude it altogether.
         * Set to a string to explicitly set the charset.
         * Defaults to true.
         */
        charset?: boolean | string;
    };
    /**
     * Incluir ou não views no dump do schema.
     * - Objeto: inclui views e aplica as opções (createOrReplace, definer, etc.).
     * - false: não inclui views no dump.
     * Defaults a objeto (views incluídas).
     */
    view?:
        | false
        | {
              createOrReplace?: boolean;
              definer?: boolean;
              algorithm?: boolean;
              sqlSecurity?: boolean;
          };
}